+++ core-service/internal/models/category.go
@@ -4,7 +4,7 @@ import "time"
 
 // tabel `categories`.
 type Category struct {
-	ID           uint64    `gorm:"column:id;primaryKey"`
+	ID           int       `gorm:"column:id;primaryKey"`
 	CategoryName string    `gorm:"column:category_name"`
 	CategorySlug string    `gorm:"column:category_slug"`
 	CreatedAt    time.Time `gorm:"column:created_at"`
