+++ core-service/internal/modules/product/product_dto.go
@@ -3,7 +3,7 @@ package product
 // GET /products.
 type ListRequest struct {
 	Keyword    string `query:"keyword"`
-	CategoryID uint64 `query:"category_id"`
+	CategoryID int    `query:"category_id"`
 	Page       int    `query:"page" validate:"omitempty,min=1"`
 	Limit      int    `query:"limit" validate:"omitempty,min=1,max=100"`
 }
@@ -19,10 +19,10 @@ func (r *ListRequest) Normalize() {
 }
 
 type ProductResponse struct {
-	ID             uint64   `json:"id"`
+	ID             int      `json:"id"`
 	ProductName    string   `json:"product_name"`
 	ProductSlug    string   `json:"product_slug"`
-	CategoryID     uint64   `json:"category_id"`
+	CategoryID     int      `json:"category_id"`
 	CategoryName   string   `json:"category_name"`
 	Unit           string   `json:"unit"`
 	Stock          int      `json:"stock"`
@@ -36,7 +36,7 @@ type ProductResponse struct {
 type ProductDetailResponse struct {
 	ProductResponse
 	Description string   `json:"description"`
-	SupplierID  uint64   `json:"supplier_id"`
+	SupplierID  int      `json:"supplier_id"`
 	Images      []string `json:"images"`
 }
 
@@ -50,7 +50,7 @@ type ListResponse struct {
 // POST /supplier/products
 type CreateProductRequest struct {
 	ProductName string  `json:"product_name" validate:"required"`
-	CategoryID  uint64  `json:"category_id" validate:"required"`
+	CategoryID  int     `json:"category_id" validate:"required"`
 	Unit        string  `json:"unit" validate:"required"`
 	Stock       int     `json:"stock" validate:"min=0"`
 	Price       float64 `json:"price" validate:"required,min=0"`
@@ -60,7 +60,7 @@ type CreateProductRequest struct {
 // PUT /supplier/products/:id
 type UpdateProductRequest struct {
 	ProductName string  `json:"product_name" validate:"required"`
-	CategoryID  uint64  `json:"category_id" validate:"required"`
+	CategoryID  int     `json:"category_id" validate:"required"`
 	Unit        string  `json:"unit" validate:"required"`
 	Price       float64 `json:"price" validate:"required,min=0"`
 	Description string  `json:"description"`
