+++ core-service/internal/modules/product/product_service_test.go
@@ -36,7 +36,7 @@ func (m *mockRepository) FindBySlug(_ context.Context, slug string) (*models.Pro
 	return nil, gorm.ErrRecordNotFound
 }
 
-func (m *mockRepository) FindByID(_ context.Context, id uint64) (*models.Product, error) {
+func (m *mockRepository) FindByID(_ context.Context, id int) (*models.Product, error) {
 	for _, p := range m.products {
 		if p.ID == id {
 			return &p, nil
@@ -48,7 +48,7 @@ func (m *mockRepository) FindByID(_ context.Context, id uint64) (*models.Product
 	return nil, gorm.ErrRecordNotFound
 }
 
-func (m *mockRepository) FindAllBySupplier(_ context.Context, supplierID uint64) ([]models.Product, error) {
+func (m *mockRepository) FindAllBySupplier(_ context.Context, supplierID int) ([]models.Product, error) {
 	var result []models.Product
 	for _, p := range m.products {
 		if p.SupplierID == supplierID {
@@ -73,7 +73,7 @@ func (m *mockRepository) Update(_ context.Context, p *models.Product) error {
 	return gorm.ErrRecordNotFound
 }
 
-func (m *mockRepository) Delete(_ context.Context, id uint64) error {
+func (m *mockRepository) Delete(_ context.Context, id int) error {
 	for i := range m.products {
 		if m.products[i].ID == id {
 			m.products = append(m.products[:i], m.products[i+1:]...)
